+++ web/components/contract/prob-change-table.tsx
@@ -9,129 +9,8 @@ import { forwardRef } from 'react'
 import { useContract } from 'web/hooks/use-contracts'
 import { Avatar } from '../widgets/avatar'
 import { Row } from '../layout/row'
-import { formatMoney, formatPercentShort } from 'common/util/format'
+import { formatPercentShort } from 'common/util/format'
 import { LoadingIndicator } from '../widgets/loading-indicator'
-
-export function ContractChangeTable(props: {
-  contractMetrics:
-    | { contract: CPMMContract; metrics: ContractMetrics }[]
-    | undefined
-}) {
-  const { contractMetrics } = props
-  console.log('contractMetrics', contractMetrics)
-
-  if (!contractMetrics) return <LoadingIndicator />
-
-  const sortedContractMetrics = sortBy(contractMetrics, ({ contract }) =>
-    Math.abs(contract.probChanges?.day ?? 0)
-  ).reverse()
-
-  if (sortedContractMetrics.length === 0)
-    return <div className="text-ink-500 px-4">None</div>
-
-  return (
-    <Col className="bg-canvas-0 divide-ink-400 border-ink-400 w-full divide-y-[0.5px] rounded-sm border-[0.5px]">
-      {sortedContractMetrics.map(({ contract, metrics }) => (
-        <ContractChangeRow
-          key={contract.id}
-          contract={contract}
-          metrics={metrics}
-        />
-      ))}
-    </Col>
-  )
-}
-const ContractChangeRow = forwardRef(
-  (
-    props: {
-      contract: CPMMContract
-      onContractClick?: (contract: Contract) => void
-      metrics: ContractMetrics
-      className?: string
-    },
-    ref: React.Ref<HTMLAnchorElement>
-  ) => {
-    const { metrics, onContractClick, className } = props
-    const contract = (useContract(props.contract.id) ??
-      props.contract) as CPMMContract
-    const {
-      creatorUsername,
-      creatorAvatarUrl,
-      closeTime,
-      isResolved,
-      question,
-      probChanges,
-    } = contract
-
-    const isClosed = closeTime && closeTime < Date.now()
-    const textColor = isClosed && !isResolved ? 'text-ink-500' : 'text-ink-900'
-
-    const probChangeToday = probChanges?.day ?? 0
-    const { from } = metrics
-    const todaysProfit = from ? from.day.profit : 0
-
-    return (
-      <Link
-        onClick={(e) => {
-          if (!onContractClick) return
-          onContractClick(contract)
-          e.preventDefault()
-        }}
-        ref={ref}
-        href={contractPath(contract)}
-        className={clsx(
-          'group flex flex-col gap-1 whitespace-nowrap px-4 py-3 lg:flex-row lg:gap-2',
-          'focus:bg-ink-300/30 lg:hover:bg-ink-300/30 transition-colors',
-          className
-        )}
-      >
-        <Avatar
-          className="hidden lg:mr-1 lg:flex"
-          username={creatorUsername}
-          avatarUrl={creatorAvatarUrl}
-          size="xs"
-        />
-        <div
-          className={clsx(
-            'break-anywhere mr-0.5 whitespace-normal font-medium lg:mr-auto',
-            textColor
-          )}
-        >
-          {question}
-        </div>
-        <Row className="gap-3">
-          <Avatar
-            className="lg:hidden"
-            username={creatorUsername}
-            avatarUrl={creatorAvatarUrl}
-            size="xs"
-          />
-          <div className="min-w-[2rem] text-right font-semibold">
-            <ContractStatusLabel contract={contract} />
-          </div>
-          <div
-            className={clsx(
-              'min-w-[2rem] text-right',
-              probChangeToday >= 0 ? 'text-teal-500' : 'text-scarlet-500'
-            )}
-          >
-            {probChangeToday >= 0 ? '+' : ''}
-            {formatPercentShort(probChangeToday)}
-          </div>
-          <div
-            className={clsx(
-              'min-w-[3.25rem] sm:text-right',
-              todaysProfit >= 0 ? 'text-teal-500' : 'text-scarlet-500'
-            )}
-          >
-            {todaysProfit >= 0 ? '+' : ''}
-            {formatMoney(todaysProfit)}
-          </div>
-        </Row>
-      </Link>
-    )
-  }
-)
 
 export function ProbChangeTable(props: {
   changes: CPMMContract[] | undefined
